requirements.txt
Django>=5.0
Pillow>=10.0
fastapi>=0.110
uvicorn>=0.29
python-dotenv>=1.0
psycopg2-binary>=2.9
dj-database-url>=2.1
PyJWT>=2.8
python-multipart>=0.0.9
google-generativeai>=0.5.0
pytest>=8.0
pytest-django>=4.8
reportlab>=4.1
openpyxl>=3.1
